Split gen_rst
New modules
- execute blocks
- save images
- write rst

review image test

Load Agg in execute_blocks.py because tests load order is alphabetic

Correct rendering of traceback in rst files

Fix orphan repetition
