+++ sphinx_gallery/gen_gallery.py
@@ -16,7 +16,8 @@ import copy
 import re
 import os
 from . import glr_path_static
-from .gen_rst import generate_dir_rst, SPHX_GLR_SIG
+from .gen_rst import generate_dir_rst
+from .write_rst import SPHX_GLR_SIG
 from .docs_resolv import embed_code_links
 from .downloads import generate_zipfiles
 
@@ -113,8 +114,6 @@ def generate_gallery_rst(app):
         for workdir in [examples_dir, gallery_dir, mod_examples_dir]:
             if not os.path.exists(workdir):
                 os.makedirs(workdir)
-        # we create an index.rst with all examples
-        fhindex = open(os.path.join(gallery_dir, 'index.rst'), 'w')
         # Here we don't use an os.walk, but we recurse only twice: flat is
         # better than nested.
         this_fhindex, this_computation_times = \
@@ -128,7 +127,10 @@ def generate_gallery_rst(app):
 
         computation_times += this_computation_times
 
-        fhindex.write(this_fhindex)
+        # we create an index.rst with all examples
+        fhindex = open(os.path.join(gallery_dir, 'index.rst'), 'w')
+        # :orphan: to suppress "not included in TOCTREE" sphinx warnings
+        fhindex.write(":orphan:\n\n" + this_fhindex)
         for directory in sorted(os.listdir(examples_dir)):
             if os.path.isdir(os.path.join(examples_dir, directory)):
                 src_dir = os.path.join(examples_dir, directory)
