+++ sphinx_gallery/gen_rst.py
@@ -14,42 +14,13 @@ Files that generate images should start with 'plot'
 """
 # Don't use unicode_literals here (be explicit with u"..." instead) otherwise
 # tricky errors come up with exec(code_blocks, ...) calls
-from __future__ import division, print_function, absolute_import
-from time import time
+from __future__ import absolute_import, division, print_function
 import codecs
 import hashlib
 import os
 import re
 import shutil
-import subprocess
-import sys
-import traceback
-import warnings
 
-
-# Try Python 2 first, otherwise load from Python 3
-try:
-    # textwrap indent only exists in python 3
-    from textwrap import indent
-except ImportError:
-    def indent(text, prefix, predicate=None):
-        """Adds 'prefix' to the beginning of selected lines in 'text'.
-
-        If 'predicate' is provided, 'prefix' will only be added to the lines
-        where 'predicate(line)' is True. If 'predicate' is not provided,
-        it will default to adding 'prefix' to all non-empty lines that do not
-        consist solely of whitespace characters.
-        """
-        if predicate is None:
-            def predicate(line):
-                return line.strip()
-
-        def prefixed_lines():
-            for line in text.splitlines(True):
-                yield (prefix + line if predicate(line) else line)
-        return ''.join(prefixed_lines())
-
-from io import StringIO
 
 try:
     # make sure that the Agg backend is set before importing any
@@ -79,91 +50,11 @@ except ImportError:
 
 from . import glr_path_static
 from .backreferences import write_backreferences, _thumbnail_div
-from .downloads import CODE_DOWNLOAD
-from .py_source_parser import (get_docstring_and_rest,
-                               split_code_and_text_blocks)
-
-from .notebook import jupyter_notebook, text2string, save_notebook
-
-try:
-    basestring
-except NameError:
-    basestring = str
-    unicode = str
-
-
-###############################################################################
-
-
-class Tee(object):
-    """A tee object to redirect streams to multiple outputs"""
-
-    def __init__(self, file1, file2):
-        self.file1 = file1
-        self.file2 = file2
-
-    def write(self, data):
-        self.file1.write(data)
-        self.file2.write(data)
-
-    def flush(self):
-        self.file1.flush()
-        self.file2.flush()
-
-    # When called from a local terminal seaborn needs it in Python3
-    def isatty(self):
-        self.file1.isatty()
-
-
-class MixedEncodingStringIO(StringIO):
-    """Helper when both ASCII and unicode strings will be written"""
-
-    def write(self, data):
-        if not isinstance(data, unicode):
-            data = data.decode('utf-8')
-        StringIO.write(self, data)
-
-
-###############################################################################
-# The following strings are used when we have several pictures: we use
-# an html div tag that our CSS uses to turn the lists into horizontal
-# lists.
-HLIST_HEADER = """
-.. rst-class:: sphx-glr-horizontal
-
-"""
-
-HLIST_IMAGE_TEMPLATE = """
-    *
-
-      .. image:: /%s
-            :scale: 47
-"""
-
-SINGLE_IMAGE = """
-.. image:: /%s
-    :align: center
-"""
-
-
-# This one could contain unicode
-CODE_OUTPUT = u""".. rst-class:: sphx-glr-script-out
-
- Out::
-
-{0}\n"""
-
-
-SPHX_GLR_SIG = """\n.. rst-class:: sphx-glr-signature
-
-    `Generated by Sphinx-Gallery <http://sphinx-gallery.readthedocs.io>`_\n"""
-
-
-def codestr2rst(codestr, lang='python'):
-    """Return reStructuredText code block from code string"""
-    code_directive = "\n.. code-block:: {0}\n\n".format(lang)
-    indented_block = indent(codestr, ' ' * 4)
-    return code_directive + indented_block
+from .py_source_parser import get_docstring_and_rest, split_code_and_text_blocks
+from .notebook import jupyter_notebook, save_notebook
+from .write_rst import rst_notebook
+from .save_images import scale_image
+from .execute_blocks import execute_script_blocks
 
 
 def extract_thumbnail_number(text):
@@ -229,141 +120,6 @@ def md5sum_is_current(src_file):
     return False
 
 
-def save_figures(image_path, fig_count, gallery_conf):
-    """Save all open matplotlib figures of the example code-block
-
-    Parameters
-    ----------
-    image_path : str
-        Path where plots are saved (format string which accepts figure number)
-    fig_count : int
-        Previous figure number count. Figure number add from this number
-    gallery_conf : dict
-        Contains the configuration of Sphinx-Gallery
-
-    Returns
-    -------
-    images_rst : str
-        rst code to embed the images in the document
-    fig_num : int
-        number of figures saved
-    """
-    figure_list = []
-
-    for fig_num in plt.get_fignums():
-        # Set the fig_num figure as the current figure as we can't
-        # save a figure that's not the current figure.
-        fig = plt.figure(fig_num)
-        kwargs = {}
-        to_rgba = matplotlib.colors.colorConverter.to_rgba
-        for attr in ['facecolor', 'edgecolor']:
-            fig_attr = getattr(fig, 'get_' + attr)()
-            default_attr = matplotlib.rcParams['figure.' + attr]
-            if to_rgba(fig_attr) != to_rgba(default_attr):
-                kwargs[attr] = fig_attr
-
-        current_fig = image_path.format(fig_count + fig_num)
-        fig.savefig(current_fig, **kwargs)
-        figure_list.append(current_fig)
-
-    if gallery_conf.get('find_mayavi_figures', False):
-        from mayavi import mlab
-        e = mlab.get_engine()
-        last_matplotlib_fig_num = fig_count + len(figure_list)
-        total_fig_num = last_matplotlib_fig_num + len(e.scenes)
-        mayavi_fig_nums = range(last_matplotlib_fig_num + 1, total_fig_num + 1)
-
-        for scene, mayavi_fig_num in zip(e.scenes, mayavi_fig_nums):
-            current_fig = image_path.format(mayavi_fig_num)
-            mlab.savefig(current_fig, figure=scene)
-            # make sure the image is not too large
-            scale_image(current_fig, current_fig, 850, 999)
-            figure_list.append(current_fig)
-        mlab.close(all=True)
-
-    return figure_rst(figure_list, gallery_conf['src_dir'])
-
-
-def figure_rst(figure_list, sources_dir):
-    """Given a list of paths to figures generate the corresponding rst
-
-    Depending on whether we have one or more figures, we use a
-    single rst call to 'image' or a horizontal list.
-
-    Parameters
-    ----------
-    figure_list : list of str
-        Strings are the figures' absolute paths
-    sources_dir : str
-        absolute path of Sphinx documentation sources
-
-    Returns
-    -------
-    images_rst : str
-        rst code to embed the images in the document
-    fig_num : int
-        number of figures saved
-    """
-
-    figure_list = [os.path.relpath(figure_path, sources_dir)
-                   for figure_path in figure_list]
-    images_rst = ""
-    if len(figure_list) == 1:
-        figure_name = figure_list[0]
-        images_rst = SINGLE_IMAGE % figure_name.lstrip('/')
-    elif len(figure_list) > 1:
-        images_rst = HLIST_HEADER
-        for figure_name in figure_list:
-            images_rst += HLIST_IMAGE_TEMPLATE % figure_name.lstrip('/')
-
-    return images_rst, len(figure_list)
-
-
-def scale_image(in_fname, out_fname, max_width, max_height):
-    """Scales an image with the same aspect ratio centered in an
-       image with a given max_width and max_height
-       if in_fname == out_fname the image can only be scaled down
-    """
-    # local import to avoid testing dependency on PIL:
-    try:
-        from PIL import Image
-    except ImportError:
-        import Image
-    img = Image.open(in_fname)
-    width_in, height_in = img.size
-    scale_w = max_width / float(width_in)
-    scale_h = max_height / float(height_in)
-
-    if height_in * scale_w <= max_height:
-        scale = scale_w
-    else:
-        scale = scale_h
-
-    if scale >= 1.0 and in_fname == out_fname:
-        return
-
-    width_sc = int(round(scale * width_in))
-    height_sc = int(round(scale * height_in))
-
-    # resize the image
-    img.thumbnail((width_sc, height_sc), Image.ANTIALIAS)
-
-    # insert centered
-    thumb = Image.new('RGB', (max_width, max_height), (255, 255, 255))
-    pos_insert = ((max_width - width_sc) // 2, (max_height - height_sc) // 2)
-    thumb.paste(img, pos_insert)
-
-    thumb.save(out_fname)
-    # Use optipng to perform lossless compression on the resized image if
-    # software is installed
-    if os.environ.get('SKLEARN_DOC_OPTIPNG', False):
-        try:
-            subprocess.call(["optipng", "-quiet", "-o", "9", out_fname])
-        except Exception:
-            warnings.warn('Install optipng to reduce the size of the \
-                          generated images')
-
-
 def save_thumbnail(image_path_template, src_file, gallery_conf):
     """Save the thumbnail image"""
     # read specification of the figure to display as thumbnail from main text
@@ -404,13 +160,6 @@ def generate_dir_rst(src_dir, target_dir, gallery_conf, seen_backrefs):
         print(80 * '_')
         return "", []  # because string is an expected return type
 
-    # suppress "not included in TOCTREE" sphinx warnings
-    fhindex = ":orphan:\n\n"
-    with open(os.path.join(src_dir, 'README.txt')) as fid:
-        fhindex += fid.read()
-    # Add empty lines to avoid bug in issue #165
-    fhindex += "\n\n"
-
     if not os.path.exists(target_dir):
         os.makedirs(target_dir)
     sorted_listdir = [fname for fname in sorted(os.listdir(src_dir))
@@ -437,6 +186,30 @@ def generate_dir_rst(src_dir, target_dir, gallery_conf, seen_backrefs):
     # sort to have the smallest entries in the beginning
     entries_text.sort()
 
+    fhindex = gallery_rst(src_dir, entries_text)
+
+    return fhindex, computation_times
+
+
+def gallery_rst(src_dir, entries_text):
+    """Generates the rst text of a gallery given the entries
+
+    Parameters
+    ----------
+    src_dir : path to the gallery sources
+    entries_text : list of 2-tuples
+        tuples contain the amount_of_code of the example
+        and the thumbnail div element
+
+    Returns:
+    str : Content of the gallery
+    """
+
+    with open(os.path.join(src_dir, 'README.txt')) as fid:
+        fhindex = fid.read()
+    # Add empty lines to avoid bug in issue #165
+    fhindex += "\n\n"
+
     for _, entry_text in entries_text:
         fhindex += entry_text
 
@@ -444,95 +217,25 @@ def generate_dir_rst(src_dir, target_dir, gallery_conf, seen_backrefs):
     fhindex += """.. raw:: html\n
     <div style='clear:both'></div>\n\n"""
 
-    return fhindex, computation_times
+    return fhindex
 
 
-def execute_code_block(code_block, example_globals,
-                       block_vars, gallery_conf):
-    """Executes the code block of the example file"""
-    time_elapsed = 0
-    stdout = ''
+def prepare_execution_env(fname, target_dir, src_dir, gallery_conf):
+    image_dir = os.path.join(target_dir, 'images')
+    if not os.path.exists(image_dir):
+        os.makedirs(image_dir)
 
-    # If example is not suitable to run, skip executing its blocks
-    if not block_vars['execute_script']:
-        return stdout, time_elapsed
+    base_image_name = os.path.splitext(fname)[0]
+    image_fname = 'sphx_glr_' + base_image_name + '_{0:03}.png'
+    image_path_template = os.path.join(image_dir, image_fname)
 
-    plt.close('all')
-    cwd = os.getcwd()
-    # Redirect output to stdout and
-    orig_stdout = sys.stdout
-    src_file = block_vars['src_file']
+    src_file = os.path.normpath(os.path.join(src_dir, fname))
+    filename_pattern = gallery_conf.get('filename_pattern')
+    execute_script = re.search(filename_pattern, src_file) and gallery_conf[
+        'plot_gallery']
 
-    try:
-        # First cd in the original example dir, so that any file
-        # created by the example get created in this directory
-        os.chdir(os.path.dirname(src_file))
-        my_buffer = MixedEncodingStringIO()
-        my_stdout = Tee(sys.stdout, my_buffer)
-        sys.stdout = my_stdout
-
-        t_start = time()
-        # don't use unicode_literals at the top of this file or you get
-        # nasty errors here on Py2.7
-        exec(code_block, example_globals)
-        time_elapsed = time() - t_start
-
-        sys.stdout = orig_stdout
-
-        my_stdout = my_buffer.getvalue().strip().expandtabs()
-        # raise RuntimeError
-        if my_stdout:
-            stdout = CODE_OUTPUT.format(indent(my_stdout, u' ' * 4))
-        os.chdir(cwd)
-        images_rst, fig_num = save_figures(block_vars['image_path'],
-                                           block_vars['fig_count'], gallery_conf)
-
-    except Exception:
-        formatted_exception = traceback.format_exc()
-
-        fail_example_warning = 80 * '_' + '\n' + \
-            '%s failed to execute correctly:' % src_file + \
-            formatted_exception + 80 * '_' + '\n'
-        warnings.warn(fail_example_warning)
-
-        fig_num = 0
-        images_rst = codestr2rst(formatted_exception, lang='pytb')
-
-        # Breaks build on first example error
-        # XXX This check can break during testing e.g. if you uncomment the
-        # `raise RuntimeError` by the `my_stdout` call, maybe use `.get()`?
-        if gallery_conf['abort_on_example_error']:
-            raise
-        # Stores failing file
-        gallery_conf['failing_examples'][src_file] = formatted_exception
-        block_vars['execute_script'] = False
-
-    finally:
-        os.chdir(cwd)
-        sys.stdout = orig_stdout
-
-    code_output = u"\n{0}\n\n{1}\n\n".format(images_rst, stdout)
-    block_vars['fig_count'] += fig_num
-
-    return code_output, time_elapsed
-
-
-def clean_modules():
-    """Remove "unload" seaborn from the name space
-
-    After a script is executed it can load a variety of setting that one
-    does not want to influence in other examples in the gallery."""
-
-    # Horrible code to 'unload' seaborn, so that it resets
-    # its default when is load
-    # Python does not support unloading of modules
-    # https://bugs.python.org/issue9072
-    for module in list(sys.modules.keys()):
-        if 'seaborn' in module:
-            del sys.modules[module]
-
-    # Reset Matplotlib to default
-    plt.rcdefaults()
+    return {'execute_script': execute_script, 'fig_count': 0,
+            'image_path': image_path_template, 'src_file': src_file}
 
 
 def generate_file_rst(fname, target_dir, src_dir, gallery_conf):
@@ -557,66 +260,12 @@ def generate_file_rst(fname, target_dir, src_dir, gallery_conf):
     if md5sum_is_current(example_file):
         return amount_of_code, 0
 
-    image_dir = os.path.join(target_dir, 'images')
-    if not os.path.exists(image_dir):
-        os.makedirs(image_dir)
-
-    base_image_name = os.path.splitext(fname)[0]
-    image_fname = 'sphx_glr_' + base_image_name + '_{0:03}.png'
-    build_image_dir = os.path.relpath(image_dir, gallery_conf['src_dir'])
-    image_path_template = os.path.join(image_dir, image_fname)
-
-    ref_fname = os.path.relpath(example_file, gallery_conf['src_dir'])
-    ref_fname = ref_fname.replace(os.path.sep, '_')
-    example_rst = """\n\n.. _sphx_glr_{0}:\n\n""".format(ref_fname)
-
-    filename_pattern = gallery_conf.get('filename_pattern')
-    execute_script = re.search(filename_pattern, src_file) and gallery_conf[
-        'plot_gallery']
-    example_globals = {
-        # A lot of examples contains 'print(__doc__)' for example in
-        # scikit-learn so that running the example prints some useful
-        # information. Because the docstring has been separated from
-        # the code blocks in sphinx-gallery, __doc__ is actually
-        # __builtin__.__doc__ in the execution context and we do not
-        # want to print it
-        '__doc__': '',
-        # Examples may contain if __name__ == '__main__' guards
-        # for in example scikit-learn if the example uses multiprocessing
-        '__name__': '__main__',
-    }
-
-    # A simple example has two blocks: one for the
-    # example introduction/explanation and one for the code
-    is_example_notebook_like = len(script_blocks) > 2
-    time_elapsed = 0
-    block_vars = {'execute_script': execute_script, 'fig_count': 0,
-                  'image_path': image_path_template, 'src_file': src_file}
+    block_vars = prepare_execution_env(
+        fname, target_dir, src_dir, gallery_conf)
     if block_vars['execute_script']:
         print('Executing file %s' % src_file)
-    for blabel, bcontent in script_blocks:
-        if blabel == 'code':
-            code_output, rtime = execute_code_block(bcontent,
-                                                    example_globals,
-                                                    block_vars,
-                                                    gallery_conf)
-
-            time_elapsed += rtime
-
-            if is_example_notebook_like:
-                example_rst += codestr2rst(bcontent) + '\n'
-                example_rst += code_output
-            else:
-                example_rst += code_output
-                if 'sphx-glr-script-out' in code_output:
-                    # Add some vertical space after output
-                    example_rst += "\n\n|\n\n"
-                example_rst += codestr2rst(bcontent) + '\n'
-
-        else:
-            example_rst += text2string(bcontent) + '\n'
-
-    clean_modules()
+    executed_blocks, time_elapsed = execute_script_blocks(
+        script_blocks, block_vars, gallery_conf)
 
     # Writes md5 checksum if example has build correctly
     # not failed and was initially meant to run(no-plot shall not cache md5sum)
@@ -624,20 +273,13 @@ def generate_file_rst(fname, target_dir, src_dir, gallery_conf):
         with open(example_file + '.md5', 'w') as file_checksum:
             file_checksum.write(get_md5sum(example_file))
 
-    save_thumbnail(image_path_template, src_file, gallery_conf)
+    save_thumbnail(block_vars["image_path"], src_file, gallery_conf)
+    write_fname = os.path.relpath(example_file, gallery_conf['src_dir'])
+    with codecs.open(example_file[:-3] + '.rst', mode='w', encoding='utf-8') as f:
+        f.write(rst_notebook(executed_blocks, write_fname, time_elapsed))
 
-    time_m, time_s = divmod(time_elapsed, 60)
-    example_nb = jupyter_notebook(script_blocks)
-    save_notebook(example_nb, example_file.replace('.py', '.ipynb'))
-    with codecs.open(os.path.join(target_dir, base_image_name + '.rst'),
-                     mode='w', encoding='utf-8') as f:
-        example_rst += "**Total running time of the script:**" \
-                       " ({0: .0f} minutes {1: .3f} seconds)\n\n".format(
-                           time_m, time_s)
-        example_rst += CODE_DOWNLOAD.format(fname,
-                                            fname.replace('.py', '.ipynb'))
-        example_rst += SPHX_GLR_SIG
-        f.write(example_rst)
+    save_notebook(jupyter_notebook(script_blocks),
+                  example_file.replace('.py', '.ipynb'))
 
     if block_vars['execute_script']:
         print("{0} ran in : {1:.2g} seconds\n".format(src_file, time_elapsed))
